requirements.txt
slackclient>=2.0.0
slackeventsapi>=2.1.0
Flask>=1.1.2
certifi~=2020.4.5.1
requests~=2.23.0
mysql-connector-python~=8.0.20
# gunicorn if building on Unix, waitress if building on Windows
gunicorn>=20.0.0
waitress~=1.4.4